moved FM index code to its own file to hide implementation details
